# stats237_course_v1_3_0/api/openapi_snapshot.json
{
  "components": {
    "schemas": {
      "AsianMCRequest": {
        "properties": {
          "K": {
            "exclusiveMinimum": 0.0,
            "title": "K",
            "type": "number"
          },
          "S0": {
            "exclusiveMinimum": 0.0,
            "title": "S0",
            "type": "number"
          },
          "T": {
            "exclusiveMinimum": 0.0,
            "title": "T",
            "type": "number"
          },
          "alpha": {
            "default": 0.05,
            "title": "Alpha",
            "type": "number"
          },
          "antithetic": {
            "default": true,
            "title": "Antithetic",
            "type": "boolean"
          },
          "method": {
            "default": "plain",
            "enum": [
              "plain",
              "lhs",
              "sobol",
              "halton"
            ],
            "title": "Method",
            "type": "string"
          },
          "n_obs": {
            "default": 50,
            "maximum": 5000.0,
            "minimum": 1.0,
            "title": "N Obs",
            "type": "integer"
          },
          "n_paths": {
            "default": 20000,
            "maximum": 5000000.0,
            "minimum": 1000.0,
            "title": "N Paths",
            "type": "integer"
          },
          "qmc_scramble": {
            "default": true,
            "title": "Qmc Scramble",
            "type": "boolean"
          },
          "r": {
            "title": "R",
            "type": "number"
          },
          "seed": {
            "default": 123,
            "title": "Seed",
            "type": "integer"
          },
          "sigma": {
            "exclusiveMinimum": 0.0,
            "title": "Sigma",
            "type": "number"
          },
          "use_control_variate": {
            "default": true,
            "title": "Use Control Variate",
            "type": "boolean"
          },
          "use_extra_control": {
            "default": true,
            "title": "Use Extra Control",
            "type": "boolean"
          }
        },
        "required": [
          "S0",
          "K",
          "r",
          "T",
          "sigma"
        ],
        "title": "AsianMCRequest",
        "type": "object"
      },
      "BSRequest": {
        "properties": {
          "K": {
            "exclusiveMinimum": 0.0,
            "title": "K",
            "type": "number"
          },
          "S0": {
            "exclusiveMinimum": 0.0,
            "title": "S0",
            "type": "number"
          },
          "T": {
            "exclusiveMinimum": 0.0,
            "title": "T",
            "type": "number"
          },
          "is_call": {
            "default": true,
            "title": "Is Call",
            "type": "boolean"
          },
          "q": {
            "default": 0.0,
            "title": "Q",
            "type": "number"
          },
          "r": {
            "title": "R",
            "type": "number"
          },
          "sigma": {
            "exclusiveMinimum": 0.0,
            "title": "Sigma",
            "type": "number"
          }
        },
        "required": [
          "S0",
          "K",
          "r",
          "T",
          "sigma"
        ],
        "title": "BSRequest",
        "type": "object"
      },
      "BasketMCRequest": {
        "properties": {
          "K": {
            "exclusiveMinimum": 0.0,
            "title": "K",
            "type": "number"
          },
          "S0": {
            "items": {
              "type": "number"
            },
            "title": "S0",
            "type": "array"
          },
          "T": {
            "exclusiveMinimum": 0.0,
            "title": "T",
            "type": "number"
          },
          "alpha": {
            "default": 0.05,
            "title": "Alpha",
            "type": "number"
          },
          "antithetic": {
            "default": true,
            "title": "Antithetic",
            "type": "boolean"
          },
          "corr": {
            "items": {
              "items": {
                "type": "number"
              },
              "type": "array"
            },
            "title": "Corr",
            "type": "array"
          },
          "lhs": {
            "default": false,
            "title": "Lhs",
            "type": "boolean"
          },
          "method": {
            "default": "plain",
            "enum": [
              "plain",
              "lhs",
              "sobol",
              "halton"
            ],
            "title": "Method",
            "type": "string"
          },
          "n_paths": {
            "default": 20000,
            "maximum": 5000000.0,
            "minimum": 1000.0,
            "title": "N Paths",
            "type": "integer"
          },
          "qmc_scramble": {
            "default": true,
            "title": "Qmc Scramble",
            "type": "boolean"
          },
          "r": {
            "title": "R",
            "type": "number"
          },
          "seed": {
            "default": 123,
            "title": "Seed",
            "type": "integer"
          },
          "use_control_variate": {
            "default": true,
            "title": "Use Control Variate",
            "type": "boolean"
          },
          "use_extra_control": {
            "default": true,
            "title": "Use Extra Control",
            "type": "boolean"
          },
          "vol": {
            "items": {
              "type": "number"
            },
            "title": "Vol",
            "type": "array"
          },
          "w": {
            "items": {
              "type": "number"
            },
            "title": "W",
            "type": "array"
          }
        },
        "required": [
          "S0",
          "w",
          "K",
          "r",
          "T",
          "vol",
          "corr"
        ],
        "title": "BasketMCRequest",
        "type": "object"
      },
      "BinomialRequest": {
        "properties": {
          "K": {
            "exclusiveMinimum": 0.0,
            "title": "K",
            "type": "number"
          },
          "S0": {
            "exclusiveMinimum": 0.0,
            "title": "S0",
            "type": "number"
          },
          "T": {
            "exclusiveMinimum": 0.0,
            "title": "T",
            "type": "number"
          },
          "exercise": {
            "default": "european",
            "enum": [
              "european",
              "american"
            ],
            "title": "Exercise",
            "type": "string"
          },
          "is_call": {
            "default": true,
            "title": "Is Call",
            "type": "boolean"
          },
          "n": {
            "default": 100,
            "maximum": 10000.0,
            "minimum": 1.0,
            "title": "N",
            "type": "integer"
          },
          "r": {
            "title": "R",
            "type": "number"
          },
          "sigma": {
            "exclusiveMinimum": 0.0,
            "title": "Sigma",
            "type": "number"
          }
        },
        "required": [
          "S0",
          "K",
          "r",
          "T",
          "sigma"
        ],
        "title": "BinomialRequest",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "IVCurveFromPricesRequest": {
        "properties": {
          "S0": {
            "exclusiveMinimum": 0.0,
            "title": "S0",
            "type": "number"
          },
          "T": {
            "exclusiveMinimum": 0.0,
            "title": "T",
            "type": "number"
          },
          "clamp_high": {
            "default": 5.0,
            "title": "Clamp High",
            "type": "number"
          },
          "clamp_low": {
            "default": 1e-06,
            "title": "Clamp Low",
            "type": "number"
          },
          "fit": {
            "default": "pchip",
            "enum": [
              "linear",
              "pchip"
            ],
            "title": "Fit",
            "type": "string"
          },
          "is_call": {
            "default": true,
            "title": "Is Call",
            "type": "boolean"
          },
          "prices": {
            "items": {
              "type": "number"
            },
            "title": "Prices",
            "type": "array"
          },
          "q": {
            "default": 0.0,
            "title": "Q",
            "type": "number"
          },
          "query_strikes": {
            "anyOf": [
              {
                "items": {
                  "type": "number"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Query Strikes"
          },
          "r": {
            "title": "R",
            "type": "number"
          },
          "strikes": {
            "items": {
              "type": "number"
            },
            "title": "Strikes",
            "type": "array"
          }
        },
        "required": [
          "strikes",
          "prices",
          "S0",
          "r",
          "T"
        ],
        "title": "IVCurveFromPricesRequest",
        "type": "object"
      },
      "IVCurveFromVolsRequest": {
        "properties": {
          "fit": {
            "default": "pchip",
            "enum": [
              "linear",
              "pchip"
            ],
            "title": "Fit",
            "type": "string"
          },
          "query_strikes": {
            "anyOf": [
              {
                "items": {
                  "type": "number"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Query Strikes"
          },
          "strikes": {
            "items": {
              "type": "number"
            },
            "title": "Strikes",
            "type": "array"
          },
          "vols": {
            "items": {
              "type": "number"
            },
            "title": "Vols",
            "type": "array"
          }
        },
        "required": [
          "strikes",
          "vols"
        ],
        "title": "IVCurveFromVolsRequest",
        "type": "object"
      },
      "IVSurfaceQueryRequest": {
        "properties": {
          "S0": {
            "exclusiveMinimum": 0.0,
            "title": "S0",
            "type": "number"
          },
          "extrapolate": {
            "default": true,
            "title": "Extrapolate",
            "type": "boolean"
          },
          "q": {
            "default": 0.0,
            "title": "Q",
            "type": "number"
          },
          "query_K": {
            "items": {
              "type": "number"
            },
            "title": "Query K",
            "type": "array"
          },
          "query_T": {
            "items": {
              "type": "number"
            },
            "title": "Query T",
            "type": "array"
          },
          "r": {
            "title": "R",
            "type": "number"
          },
          "smiles": {
            "items": {
              "$ref": "#/components/schemas/SurfaceSmileSlice"
            },
            "title": "Smiles",
            "type": "array"
          }
        },
        "required": [
          "smiles",
          "S0",
          "r",
          "query_T",
          "query_K"
        ],
        "title": "IVSurfaceQueryRequest",
        "type": "object"
      },
      "ImpliedVolRequest": {
        "properties": {
          "K": {
            "exclusiveMinimum": 0.0,
            "title": "K",
            "type": "number"
          },
          "S0": {
            "exclusiveMinimum": 0.0,
            "title": "S0",
            "type": "number"
          },
          "T": {
            "exclusiveMinimum": 0.0,
            "title": "T",
            "type": "number"
          },
          "is_call": {
            "default": true,
            "title": "Is Call",
            "type": "boolean"
          },
          "max_iter": {
            "default": 200,
            "title": "Max Iter",
            "type": "integer"
          },
          "price": {
            "exclusiveMinimum": 0.0,
            "title": "Price",
            "type": "number"
          },
          "q": {
            "default": 0.0,
            "title": "Q",
            "type": "number"
          },
          "r": {
            "title": "R",
            "type": "number"
          },
          "tol": {
            "default": 1e-10,
            "title": "Tol",
            "type": "number"
          },
          "vol_high": {
            "default": 5.0,
            "title": "Vol High",
            "type": "number"
          },
          "vol_low": {
            "default": 1e-06,
            "title": "Vol Low",
            "type": "number"
          }
        },
        "required": [
          "price",
          "S0",
          "K",
          "r",
          "T"
        ],
        "title": "ImpliedVolRequest",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Stats237 Quant API",
    "version": "1.3.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/calibration/iv_curve/from_prices": {
      "post": {
        "operationId": "calibration_iv_curve_from_prices_calibration_iv_curve_from_prices_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-Request-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Request-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IVCurveFromPricesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Calibration Iv Curve From Prices Calibration Iv Curve From Prices Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Calibration Iv Curve From Prices"
      }
    },
    "/calibration/iv_curve/from_vols": {
      "post": {
        "operationId": "calibration_iv_curve_from_vols_calibration_iv_curve_from_vols_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-Request-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Request-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IVCurveFromVolsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Calibration Iv Curve From Vols Calibration Iv Curve From Vols Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Calibration Iv Curve From Vols"
      }
    },
    "/calibration/iv_surface/query": {
      "post": {
        "operationId": "calibration_iv_surface_query_calibration_iv_surface_query_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-Request-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Request-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IVSurfaceQueryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Calibration Iv Surface Query Calibration Iv Surface Query Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Calibration Iv Surface Query"
      }
    },
    "/greeks/black_scholes": {
      "post": {
        "operationId": "greeks_black_scholes_greeks_black_scholes_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-Request-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Request-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BSRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Greeks Black Scholes Greeks Black Scholes Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Greeks Black Scholes"
      }
    },
...
    "/implied_vol": {
      "post": {
        "operationId": "implied_vol_endpoint_implied_vol_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-Request-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Request-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImpliedVolRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Implied Vol Endpoint Implied Vol Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Implied Vol Endpoint"
      }
    },
    "/mc/asian/arithmetic_call": {
      "post": {
        "operationId": "mc_asian_arithmetic_call_mc_asian_arithmetic_call_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-Request-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Request-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AsianMCRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Mc Asian Arithmetic Call Mc Asian Arithmetic Call Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Mc Asian Arithmetic Call"
      }
    },
    "/mc/basket/call": {
      "post": {
        "operationId": "mc_basket_call_mc_basket_call_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-Request-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Request-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BasketMCRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Mc Basket Call Mc Basket Call Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Mc Basket Call"
      }
    },
...
    "/price/binomial": {
      "post": {
        "operationId": "price_binomial_price_binomial_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-Request-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Request-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BinomialRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Price Binomial Price Binomial Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Price Binomial"
      }
    },
    "/price/black_scholes": {
      "post": {
        "operationId": "price_black_scholes_price_black_scholes_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-Request-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Request-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BSRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Price Black Scholes Price Black Scholes Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Price Black Scholes"
      }
    }
  }
}